Rewrite web testing
