+++ tests/web/questions.spec.ts-snapshots/element-negative-chromium-linux.png
@@ -1,0 +1,1 @@
+./element-negative-chromium-darwin.png
