+++ tests/web/questions.spec.ts-snapshots/element-positive-chromium-linux.png
@@ -1,0 +1,1 @@
+./element-positive-chromium-darwin.png
